fix : 변수형 변경

+++ CampingUsProject-customer/app/src/main/java/com/test/campingusproject_customer/dataclassmodel/ReviewModel.kt
@@ -4,7 +4,7 @@ data class ReviewModel(
     var reviewId: Long,                 // 리뷰 ID
     var reviewProductId: Long,          // 상품 ID
     var reviewWriterId: String,         // 리뷰 작성자 ID
-    var reviewRating: Float,            // 리뷰 별점
+    var reviewRating: Long,            // 리뷰 별점
     var reviewImage: String,            // 리뷰 사진
     var reviewContent: String           // 리뷰 내용
 )
